+++ src/TaggerCommand.php
@@ -9,8 +9,8 @@ declare(strict_types=1);
 
 namespace Phly\KeepAChangelog;
 
+use Phly\KeepAChangelog\Provider\GetProviderTrait;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
@@ -18,7 +18,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class TaggerCommand extends Command
 {
-    use GetChangelogFileTrait;
+    use GetChangelogFileTrait, GetProviderTrait;
 
     private const HELP = <<< 'EOH'
 Create a new git tag for the current repository, using the relevant changelog entry.
@@ -57,6 +57,12 @@ EOH;
             InputOption::VALUE_REQUIRED,
             'Alternate git tag name to use when tagging; defaults to <version>'
         );
+        $this->addOption(
+            'provider',
+            null,
+            InputOption::VALUE_OPTIONAL,
+            'Repository provider. Options: github, gitlab; defaults to github'
+        );
     }
 
     protected function execute(InputInterface $input, OutputInterface $output) : int
@@ -81,7 +87,7 @@ EOH;
         $formatter = new ChangelogFormatter();
         $changelog = $formatter->format($changelog);
 
-        if (! $this->tagWithChangelog($tagName, $package, $version, $changelog)) {
+        if (! $this->getProvider($input)->createLocalTag($tagName, $package, $version, $changelog)) {
             $output->writeln('<error>Error creating tag!</error>');
             $output->writeln('Check the output logs for details');
             return 1;
@@ -96,18 +102,5 @@ EOH;
         $output->write($changelog);
 
         return 0;
-    }
-
-    private function tagWithChangelog(string $tagName, string $package, string $version, string $changelog) : bool
-    {
-        $tempFile = tempnam(sys_get_temp_dir(), 'KAC');
-        file_put_contents($tempFile, sprintf("%s %s\n\n%s", $package, $version, $changelog));
-
-        $command = sprintf('git tag -s -F %s %s', $tempFile, $tagName);
-        system($command, $return);
-
-        unlink($tempFile);
-
-        return 0 === $return;
     }
 }